[REFAC] Simplify RestockBoat component and remove unused props; streamline restock logic in Water component
