+++ src/features/game/expansion/components/RestockBoat.tsx
@@ -1,10 +1,17 @@
 import React, { useContext, useState } from "react";
 import boat from "assets/decorations/restock_boat.png";
-import { MapPlacement } from "./MapPlacement";
-import { PIXEL_SCALE, StockableName } from "features/game/lib/constants";
+import { Coordinates, MapPlacement } from "./MapPlacement";
+import {
+  INITIAL_STOCK,
+  PIXEL_SCALE,
+  StockableName,
+} from "features/game/lib/constants";
 import { useSelector } from "@xstate/react";
 import { MachineState } from "features/game/lib/gameMachine";
-import { canRestockShipment } from "features/game/events/landExpansion/shipmentRestocked";
+import {
+  canRestockShipment,
+  SHIPMENT_STOCK,
+} from "features/game/events/landExpansion/shipmentRestocked";
 import { Context } from "features/game/GameProvider";
 import { Modal } from "components/ui/Modal";
 import { Label } from "components/ui/Label";
@@ -17,6 +24,8 @@ import { Box } from "components/ui/Box";
 import { ITEM_DETAILS } from "features/game/types/images";
 import { CROP_LIFECYCLE } from "features/island/plots/lib/plant";
 import Decimal from "decimal.js-light";
+import { SEEDS } from "features/game/types/seeds";
+import { WORKBENCH_TOOLS, TREASURE_TOOLS } from "features/game/types/tools";
 
 const expansions = (state: MachineState) =>
   state.context.state.inventory["Basic Land"]?.toNumber() ?? 3;
@@ -24,21 +33,49 @@ const canRestock = (state: MachineState) =>
   canRestockShipment({ game: state.context.state }) &&
   !!state.context.state.shipments.restockedAt;
 
-export const RestockBoat: React.FC<{
-  restockSeeds: [string, number][];
-  restockTools: [string, number][];
-  getShipmentAmount: (item: StockableName, amount: number) => Decimal;
-}> = ({ restockSeeds, restockTools, getShipmentAmount }) => {
+const _state = (state: MachineState) => state.context.state;
+
+export const RestockBoat: React.FC = () => {
   const { t } = useAppTranslation();
   const [isOpen, setIsOpen] = useState(false);
   const { gameService, showAnimations } = useContext(Context);
 
   const expansionCount = useSelector(gameService, expansions);
+  const state = useSelector(gameService, _state);
   const showShip = useSelector(gameService, canRestock);
-
   if (!showShip) return null;
+  const getShipmentAmount = (item: StockableName, amount: number): Decimal => {
+    const totalStock = INITIAL_STOCK(state)[item];
+    const remainingStock = state.stock[item] ?? new Decimal(0);
+    // If shipment amount will exceed total stock
+    if (remainingStock.add(amount).gt(totalStock)) {
+      // return the difference between total and remaining stock
+      return totalStock.sub(remainingStock);
+    } else {
+      // else return shipment stock
+      return new Decimal(amount);
+    }
+  };
 
-  const wharfCoords = () => {
+  const restockTools = Object.entries(SHIPMENT_STOCK)
+    .filter((item) => item[0] in { ...WORKBENCH_TOOLS, ...TREASURE_TOOLS })
+    .filter(([item, amount]) => {
+      const shipmentAmount = getShipmentAmount(item as StockableName, amount);
+      return shipmentAmount.gt(0);
+    });
+
+  const restockSeeds = Object.entries(SHIPMENT_STOCK)
+    .filter((item) => item[0] in SEEDS)
+    .filter(([item, amount]) => {
+      const shipmentAmount = getShipmentAmount(item as StockableName, amount);
+      return shipmentAmount.gt(0);
+    });
+
+  const restockIsEmpty = [...restockSeeds, ...restockTools].length <= 0;
+
+  if (restockIsEmpty) return null;
+
+  const wharfCoords = (): Coordinates => {
     if (expansionCount < 7) {
       return { x: -1, y: -7 };
     }
@@ -51,7 +88,7 @@ export const RestockBoat: React.FC<{
 
   return (
     <>
-      <MapPlacement x={wharfCoords().x} y={wharfCoords().y}>
+      <MapPlacement {...wharfCoords()}>
         <img
           src={boat}
           style={{
